// legacy-app/src/interfaces/drogal/get-product.api.interface.ts
export interface DrogalCustomData {
  programCode: number;
  requiresMedicalPrescription: boolean;
  pbmPrice: number;
  authVoucher: number;
  ean: number;
}

export interface DrogalApiGetProductApiResponse {
  productId: string;
  productName: string;
  brand: string;
  brandId: number;
  brandImageUrl: any;
  linkText: string;
  productReference: string;
  productReferenceCode: string;
  categoryId: string;
  productTitle: string;
  metaTagDescription: string;
  releaseDate: string;
  clusterHighlights: ClusterHighlights;
  categories: string[];
  categoriesIds: string[];
  link: string;
  'Descrição Marketplace': string[];
  Indicação: string[];
  BloqueioVenda: string[];
  Contraindicação: string[];
  'Reações Adversas': string[];
  'Como Usar': string[];
  SimilarFoto: string[];
  Bula: string[];
  'Parte do Corpo': string[];
  'Especificacoes Gerais': string[];
  allSpecifications: string[];
  allSpecificationsGroups: string[];
  description: string;
  items: Item[];
  CustomData?: string[];
  VAN?: string[];
  PBM?: string[];
}

export interface ClusterHighlights {}

export interface Item {
  itemId: string;
  name: string;
  nameComplete: string;
  complementName: string;
  ean: string;
  referenceId: ReferenceId[];
  measurementUnit: string;
  unitMultiplier: number;
  modalType: string;
  isKit: boolean;
  images: Image[];
  sellers: Seller[];
  Videos: any[];
  estimatedDateArrival: any;
}

export interface ReferenceId {
  Key: string;
  Value: string;
}

export interface Image {
  imageId: string;
  imageLabel: string;
  imageTag: string;
  imageUrl: string;
  imageText: string;
  imageLastModified: string;
}

export interface Seller {
  sellerId: string;
  sellerName: string;
  addToCartLink: string;
  sellerDefault: boolean;
  commertialOffer: CommertialOffer;
}